+++ NEWS.md
@@ -1,5 +1,8 @@
 # rsconnect (development version)
 
+* Added support for interaction with Posit Connect deployments
+  hosted in Snowpark Container Services.
+
 * Content directories with a period in their name are no longer treated as a
   document path when computing the location for deployment records. (#1138)